requirements.txt
fastapi==0.65.2
uvicorn==0.15.0
tensorflow==2.9.1
tesseract==0.1.3